# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Bookstore API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/laporan-pembayarans": {
            "get": {
                "tags": [
                    "Laporan Pembayaran"
                ],
                "summary": "List semua laporan pembayaran",
                "operationId": "33068c3f922df755c25c920e114bc34c",
                "responses": {
                    "200": {
                        "description": "Data laporan pembayaran berhasil ditampilkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LaporanPembayaran"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Laporan Pembayaran"
                ],
                "summary": "Buat laporan pembayaran baru",
                "operationId": "3d51e83f631a692b96c7981d1e410b35",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "periode_laporan",
                                    "jumlah_karyawan",
                                    "total_pengeluaran",
                                    "rata_rata"
                                ],
                                "properties": {
                                    "periode_laporan": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-01"
                                    },
                                    "jumlah_karyawan": {
                                        "type": "string",
                                        "example": "10"
                                    },
                                    "total_pengeluaran": {
                                        "type": "string",
                                        "example": "15000000"
                                    },
                                    "rata_rata": {
                                        "type": "string",
                                        "example": "1500000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil membuat laporan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LaporanPembayaran"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/laporan-pembayarans/{laporanPembayaran}": {
            "get": {
                "tags": [
                    "Laporan Pembayaran"
                ],
                "summary": "Detail laporan pembayaran",
                "operationId": "b769a7fbdfac80471d876d4bb08254ad",
                "parameters": [
                    {
                        "name": "laporanPembayaran",
                        "in": "path",
                        "description": "ID laporan pembayaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail laporan ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LaporanPembayaran"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Laporan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Laporan Pembayaran"
                ],
                "summary": "Update laporan pembayaran",
                "operationId": "aac6bdb04ed3b3a5a35c5030d817d325",
                "parameters": [
                    {
                        "name": "laporanPembayaran",
                        "in": "path",
                        "description": "ID laporan pembayaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "periode_laporan": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-01"
                                    },
                                    "jumlah_karyawan": {
                                        "type": "string",
                                        "example": "12"
                                    },
                                    "total_pengeluaran": {
                                        "type": "string",
                                        "example": "18000000"
                                    },
                                    "rata_rata": {
                                        "type": "string",
                                        "example": "1500000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Laporan berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LaporanPembayaran"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Laporan Pembayaran"
                ],
                "summary": "Hapus laporan pembayaran",
                "operationId": "b7c300ed02f1248c45c31508df954deb",
                "parameters": [
                    {
                        "name": "laporanPembayaran",
                        "in": "path",
                        "description": "ID laporan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Laporan berhasil dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/laporan-pembayarans/periode/{periode}": {
            "get": {
                "tags": [
                    "Laporan Pembayaran"
                ],
                "summary": "Ambil laporan berdasarkan periode (YYYY-MM)",
                "operationId": "1da8401af8ef1f70973b2542e2ccf972",
                "parameters": [
                    {
                        "name": "periode",
                        "in": "path",
                        "description": "Periode format YYYY-MM",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "2025-06"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Laporan berdasarkan periode",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LaporanPembayaran"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Format periode tidak valid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/laporan-pembayarans/total-pengeluaran/{total}": {
            "get": {
                "tags": [
                    "Laporan Pembayaran"
                ],
                "summary": "Ambil laporan berdasarkan total pengeluaran",
                "operationId": "07fedfd2c41b5cc6f579ceb622c95ef4",
                "parameters": [
                    {
                        "name": "total",
                        "in": "path",
                        "description": "Total pengeluaran",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "15000000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Laporan berdasarkan total pengeluaran",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LaporanPembayaran"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LaporanPembayaran": {
                "required": [
                    "periode_Laporan",
                    "jumlah_karyawan",
                    "total_pengeluaran",
                    "rata_rata"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "periode_Laporan": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-04-01"
                    },
                    "jumlah_karyawan": {
                        "type": "string",
                        "example": "10"
                    },
                    "total_pengeluaran": {
                        "type": "string",
                        "example": "15000000"
                    },
                    "rata_rata": {
                        "type": "string",
                        "example": "1500000"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-03T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-03T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication using Laravel Sanctum",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "security": [
                {
                    "sanctum": []
                }
            ]
        }
    },
...
}